Typoes fixed up. Finally using a std::map for the alphabet and also looking through it for my indexes instead of searching through the entire string each and every time I am trying to find something.

+++ itp232/stock_trader.cpp
@@ -114,7 +114,7 @@ int main(){
 	new Item("Chocolate Chip Cookies Futures",5.31), new Item("Giant Fishstick",13.31), new Item("Signed Photo of Nessie", 3.50)};
 	
 	//get the total items.
-	int total_items=items.size();
+	long long total_items=items.size();
 	
 	//prompt for name.
 	std::cout << "Please enter your name: ";
